Split LinearTrackPropagator from implementation file
